Add config files to main
