+++ src/Components/App/App.js
@@ -2,7 +2,7 @@ import './App.css';
 import Header from '../Header/Header';
 import LandingPage from '../LandingPage/LandingPage';
 import Footer from '../Footer/Footer';
-import OrderCardContainer from '../OrderContainer/OrderCardContainer';
+import OrderCardContainer from '../OrderCardContainer/OrderCardContainer';
 import Profile from '../Profile/Profile';
 import LoginButton from '../LoginButton/LoginButton';
 import LogoutButton from '../LogoutButton/LogoutButton';
