+++ src/Components/OrderCardContainer/OrderCardContainer.css
@@ -1,0 +1,13 @@
+button {
+  display: flex;
+  align-items: center;
+  background-color: #d23121;
+  color: white;
+  font-size: 24px;
+  font-weight: bold;
+  padding: 10px;
+}
+
+h1 {
+  font-size: 40px;
+}
