--- src/Components/OrderContainer/OrderCardContainer.js
@@ -1,34 +1,0 @@
-import './OrderCardContainer.css';
-
-const OrderCardContainer = () => {
-  return (
-    <>
-      <div>
-        <h1>Cheese Pizza</h1>
-        <button>Order Now</button>
-      </div>
-      <div>
-        <h1>Pepperoni Pizza</h1>
-        <button>Order Now</button>
-      </div>
-      <div>
-        <h1>Sausage Pizza</h1>
-        <button>Order Now</button>
-      </div>
-      <div>
-        <h1>Supreme Pizza</h1>
-        <button>Order Now</button>
-      </div>
-      <div>
-        <h1>Buffalo Chicken Pizza</h1>
-        <button>Order Now</button>
-      </div>
-      <div>
-        <h1>Vegetarian Pizza</h1>
-        <button>Order Now</button>
-      </div>
-    </>
-  );
-};
-
-export default OrderCardContainer;
