+++ src/auth_config.json
@@ -1,0 +1,5 @@
+{
+  "domain": "dev-3m7cus37.us.auth0.com",
+  "clientId": "eJPH4Gye0ZRi0VEWBqrSyjghHj8oz3Jd",
+  "audience": "https://dev-3m7cus37.us.auth0.com/api/v2/"
+}
